refactor: Add request and response packages in dto (#45)
Added request and response packages in dto.
