+++ BE/src/main/java/com/codesquad/team12/baseball/service/TeamService.java
@@ -1,6 +1,6 @@
 package com.codesquad.team12.baseball.service;
 
-import com.codesquad.team12.baseball.dto.TeamPlayerDto;
+import com.codesquad.team12.baseball.dto.response.TeamPlayerDto;
 import com.codesquad.team12.baseball.model.Team;
 import com.codesquad.team12.baseball.repository.TeamRepository;
 import org.springframework.stereotype.Service;
